Updated name of the shell installer and fixed typo
The shell script file's name has been changed to install.sh, so updated it in the readme file. Alternately, there was a typo in the spelling of requirements, so fixed that. The typo fixing would also require the name of the requirements.txt file to be named correctly in order to not run into problems and congestions.